# openapi.yaml
openapi: 3.0.0
servers:
  - url: http://localhost8080
info:
  description: Contract API For The Skill-Matrix Application. This API contains all the neccessary endpoints for CRUD operations based on project requirements. In a nutshell, the application for which this API was created, contains all the active employees currently working for SoftwareONE. The purpose of this application is to aggregate all employees in one place so that any manager can access and contact any employee based on skills and other criterias.
  version: "1.0.0"
  title: Skill-Matrix API 
  contact:
    email: zourab.lavrentiadis@softwareone.com
  license:
    name: SoftwareONE
    url: 'http://www.softwareone.com'
...
paths:
...
  /user:
    get:
      tags:
        - user
      summary: fetch all users from the database
      operationId: getUsers
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        4XX:
          description: problem caused by the user
        5XX:
          description: problem caused by the server   
    
    post:
      tags:
        - user
      summary: add a user to the database
      operationId: addUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        201:
           description: created
           content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        4XX:
          description: problem caused by the user
        5XX:
          description: problem caused by the server   
  /user/{userId}/skill:
    get:
      tags:
        - user
      summary: fetch skills of a single user from the database
      operationId: getUserSkills
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/userSkillInputArray'
        4XX:
          description: problem caused by the user
        5XX:
          description: problem caused by the server   
          
    post:
      tags:
        - user
      summary: add one or multiple skills to the user
      operationId: addUserSkill
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/userSkillInputArray'
      responses:
        201:
          description: created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/userSkillInputArray'
        4XX:
          description: problem caused by the user
        5XX:
          description: problem caused by the server   
    
    put:
      tags:
        - user
      summary: update one or multiple skills of a user
      operationId: updateUserSkill
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/userSkillInputArray'
      responses:
        201:
          description: created
        4XX:
          description: problem caused by the user
        5XX:
          description: problem caused by the server  
  /user/{userId}/skill/{skillId}:
    delete:
      tags:
        - user
      summary: delete a user skill by id
      operationId: deleteUserSkill
      parameters:
      - in: path
        name: userId 
        required: true
        schema:
          type: integer
      - in: path
        name: skillId
        required: true
        schema:
          type: integer
      responses:
        201:
          description: deleted
        4XX:
          description: problem caused by the user
        5XX:
          description: problem caused by the server        
  /user/{userId}/language:
    get:
      tags:
        - user
      summary: fetch languages of a single user from the database
      operationId: getUserLanguages
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/userLanguageInputArray'
        4XX:
          description: problem caused by the user
        5XX:
          description: problem caused by the server   
          
    post:
      tags:
        - user
      summary: add one or multiple languages to the user
      operationId: addUserLanguage
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/userLanguageInputArray'
      responses:
        201:
          description: created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/userLanguageInputArray'
        4XX:
          description: problem caused by the user
        5XX:
          description: problem caused by the server   
    
    put:
      tags:
        - user
      summary: update one or multiple languages of a user
      operationId: updateUserLanguage
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LanguageInput'
      responses:
        201:
          description: created
        4XX:
          description: problem caused by the user
        5XX:
          description: problem caused by the server  
  /user/{userId}/language/{languageId}:
    delete:
      tags:
        - user
      summary: delete a user language by id
      operationId: deleteUserLanguage
      parameters:
      - in: path
        name: userId 
        required: true
        schema:
          type: integer
      - in: path
        name: languageId
        required: true
        schema:
          type: integer
      responses:
        201:
          description: deleted
        4XX:
          description: problem caused by the user
        5XX:
          description: problem caused by the server 
  /user/{userId}/personalInfo:
    get:
      tags:
        - user
      summary: fetch personal information of a user from the database
      operationId: getUserPersonalInfo
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/userPersonalInfo'
        4XX:
          description: problem caused by the user
        5XX:
          description: problem caused by the server   
          
    post:
      tags:
        - user
      summary: add personal information to a single user
      operationId: addUserPersonalInfo
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/userPersonalInfo'
      responses:
        201:
          description: created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/userPersonalInfo'
        4XX:
          description: problem caused by the user
        5XX:
          description: problem caused by the server   
    
    put:
      tags:
        - user
      summary: update user's personal information
      operationId: updateUserPersonalInfo
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/userPersonalInfo'
      responses:
        201:
          description: created
        4XX:
          description: problem caused by the user
        5XX:
          description: problem caused by the server  
  /user/{userId}/personalInfo/{personalInfoId}:
    delete:
      tags:
        - user
      summary: delete a user's personal info by id
      operationId: deleteUserPersonalInfo
      parameters:
      - in: path
        name: userId 
        required: true
        schema:
          type: integer
      - in: path
        name: personalInfoId
        required: true
        schema:
          type: integer
      responses:
        201:
          description: deleted
        4XX:
          description: problem caused by the user
        5XX:
          description: problem caused by the server 
  /user/{userId}:
    get:
      tags:
        - user
      summary: find user by id
      operationId: getUserById
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        4XX:
          description: problem caused by the user
        5XX:
          description: problem caused by the server   
  
    put:
      tags:
        - user
      summary: update a user by id
      operationId: updateUser
      parameters:
      - in: path
        name: userId
        required: true
        schema:
          type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
               $ref: '#/components/schemas/User'
      responses:
        201:
          description: created
        4XX:
          description: problem caused by the user
        5XX:
          description: problem caused by the server  
          
    delete:
      tags:
        - user
      summary: delete a user by id
      operationId: deleteUser
      parameters:
      - in: path
        name: userId
        required: true
        schema:
          type: integer
      responses:
        201:
          description: deleted
        4XX:
          description: problem caused by the user
        5XX:
          description: problem caused by the server  
...
components:
  schemas:
    User:
      type: object
      required:
        - id
        - firstName
        - lastName
        - email
        - mobileNumber
        - location
        - aboutMe
        - birthDate
        - jobTitle
        - department
        - skills
        - languages
      properties:
        userId:
          type: integer
          example: 1
        firstName:
          type: string
          example: elon 
        lastName:
          type: string
          example: musk
        email:
          type: string
          example: elon-musk@gmail.com
        mobileNumber:
          type: integer
          example: 017621505142
        location: 
          type: string
          example: leipzig
        aboutMe:
          type: string
          example: java developer from Leipzig
        birthDate:
          type: string
          format: date-time
          example: '2016-08-29'
        skills:
          type: array
          items:
           $ref: '#/components/schemas/UserSkill'
        languages:
          type: array
          items:
           $ref: '#/components/schemas/UserLanguage'
    userSkillInputArray:
        type: array
        items:
            $ref: '#/components/schemas/skillInput'
...
    LanguageInput:
        type: object
        properties:
          languageId:
            type: integer
            example: 1
          languageLevel:
            type: integer
            example: 1
    userLanguageInputArray:
      type: array
      items:
        $ref: '#/components/schemas/LanguageInput'
...